Refactor model package layout
