+++ model/descriminators/__init__.py
@@ -1,0 +1,9 @@
+"""Discriminator architectures for SRGAN."""
+
+from .srgan_discriminator import Discriminator
+from .patchgan import PatchGANDiscriminator
+
+__all__ = [
+    "Discriminator",
+    "PatchGANDiscriminator",
+]
